Set up basic thermostat via new Shelly 1
The Shelly is connected in parallel with the old basic
digital thermostat, so both work independently.
keep_alive is disabled and there is a timeout set on the Shelly.
So hopefully that should prevent any disaster if Home Assistant
dies or if it gets stuck in the "on" state for the boiler.