Use 1-wire sensor directly and other cleanup
I will no longer use a filter to smooth out the temperature.
The reason I set it up in the first place was that the temp
graph looked ugly. Ever since I switched to
custom:mini-graph-card this hasn't been an issue anymore.
Also, the new sensor seems to be better in this regard - it seems
there is no up and down flactuation.