Disable ios integration, remove default values for thermostat
Default values mean that in case of a power drop out, when
HA starts again, the thermostat will suddenly be set differently.
But without keep_alive enabled it may be different than reality.
So it's better to keep the previous values and not have default
ones.

This will also export the thermostat to Homekit.