Set up zigbee2mqtt
I connected a usb zigbee dongle that I got from a colleague.
Then I installed the zigbee2mqtt addon. This was actually
quite simple to configure according to documentation.

Finally I made the changes in this commit - this will add a new
Lovelave card with a button to enable pairing of new devices
with a timer of 10 minutes.