Turn the lights off everyday at 4 a.m.
This will just add an additional trigger to the existing
automation for when I leave home. The use case for this is
that if for whatever reason the lights stay on while I'm gone
long term, they will eventually turn off. This can happen
for example after a power failure.