Improve volume controll via IKEA Symfonisk and add yamaha on/off
The problem with the symfonisk button is that when
you start rotating to one side (to increase/decrease volume),
it will only send one event when you start and then one when
you stop. So if you simply listen to the start events and then
rotate quickly, you will only change it by one step and then
nothing. If you rotate slowly, you will get the stop events
in between.

So I implemented two scripts for up/down which will incrementally
add/decrease volume in a loop with a small delay. When we get
the rotation start event, we start this loop. When we get any other
event, we stop the loop. It still doesn't work 100 % but it's
better than before. See also here for discussion about this problem:
https://github.com/Koenkk/zigbee-herdsman-converters/issues/658

On top of this, I added an automation to turn the Yamaha receiver
on/off on the Symfonisk button click.