Update ui-lovelace.yaml to latest version
I switched to UI lovelace config. Apparently
the config is in fact no longer stored as
yaml on the disk, but it's a json stored
inside .storage in the config folder. And
the yaml is automatically generated when you
go to the Raw config editor via UI.
So I can just copy the current version
to ui-lovelace.yaml manually from time
to time to have it in git.