Set up a template cover to control all blinds at once
I tried to use a cover group as well (commented out). That would
be the preferred way to do it, but in this case timing is important
and with a cover group essentially 3 mqtt messages are published
at various times, so then each blind can be in a different position
depending on exactly when it gets the commands.

Using the group topic is more reliable, but of course it's still
not 100 % - there is just one message for all blinds, but there
can still be a difference in when exactly the blind receives
the message and reacts.

Note that the position_template basically says that the blind is
always considered to be at 50 % up/down. That's because without
it HA makes its own predictions and sometimes it assumes it's
at the limit when it's not and wouldn't allow you to move
in that direction otherwise.