Set up virtual switch for presence detection
Based on https://community.home-assistant.io/t/use-any-switch-as-person-presence/207304/3

Basically all I wanted is a virtual switch that would
act as a device tracker. This switch is then added to homekit.
And on my iPhone I set up automation to turn the switch
on when I get home and off when I leave home.

One small caveat to remember: When you set up the mqtt
device tracker, it may not initially be visible in HA.
This happens if track new devices is set to false.
Find the device in known_devices.yaml and set the flag
to true.