Improve TV On and add good morning automation
TV On: Use two "choose" sections, so that we can have if conditions
in the middle of the action without ending execution.
We set volume to 0.5 if it's lower than that. We also
close the blind if it's before lunch.

Good Morning: The new automation is triggered by a button and
it will open blinds and turn on Kitchen Unit lights.