Add scripts for blinds base position and step up/down
Until now, I could only do all the way up/down
and then stop. So the set the basic position in
the afternoon on a sunny day, I would have to first
go all the way down and then start going up and
stop it manually. Now I have a script for this.

And also two other scripts to step up/down.