Improve Nightstand Switch On automation
The automation started misbehaving. It was supposed
to turn off all lights and then turn on bedroom
lights. But the service calls are asynchronous
so they're not guaranteed to be processed sequentially.

So recently the automation's result was often "all lights off".
The solution is to only turn off lights in the Living Room.