Switch to using tilt for blinds control
I updated the tasmota fw in my sonoff blind switches
to v12 and enabled tilt support. So now I'm also
updating the scripts to control the blinds to use this.

Hopefully this will make it a bit more deterministic because
using the backlog command often resulted in the blinds getting
out of sync. This is because Tasmota is not very precise in
the timing. It's still not perfect even with the tilt feature
but it seems to be a bit better.

This brings an issue though - I can no longer easily shut
the blinds at night just by using the shutterclose command
because it may already be closed, but with an open tilt.
So I will have to figure this out later - probably by writing
another script to ensure both shutter and tilt are closed.