Remove brightness_pct from scene
I found this warning in the log:
The use of other attributes than device state attributes is deprecated and will be removed in a future release. Invalid attributes are ['brightness_pct'].

So apparently you're supposed to use brightness instead - that's
the actual state attribute. brightness_pct is allowed for lights
(and is then translated into a brightness value 0-255), but for
scenes it shoulnd't be used.