Share Yamaha to Homekit via accessory mode
Previosly, the only option to share a media_player
was to expose a bunch of switch. HA 2023.10 now supports
exposing media players as receivers in HomeKit.
For that, a new integration using accessory mode
needs to be set up for that entity. This is now done
via UI. (The entity needs the device_class receiver.)