Improve setting blinds to base position
The script was waiting for all the blinds to be totally
closed before changing the tilt. But one of them
is a taller window and so there would be a period
where the other two blinds wait for the third one
and the room is very dark.

Instead, let's run sequences in parallel so that each
blind can have its tilt changed as soon as it's fully
closed.